requirements.txt
numpy
requests
bitarray
python-socketio
eventlet
